// src/@v2/documents/domain/enums/document-section-type.enum.ts
export enum DocumentSectionTypeEnum {
    CHAPTER = 'CHAPTER',
    TOC = 'TOC',
    COVER = 'COVER',
    SECTION = 'SECTION',
    APR = 'APR',
    APR_GROUP = 'APR_GROUP',
    ACTION_PLAN = 'ACTION_PLAN',
    ITERABLE_ENVIRONMENTS = 'ITERABLE_ENVIRONMENTS',
    ITERABLE_CHARACTERIZATION = 'ITERABLE_CHARACTERIZATION'
}